Add optional chaining to string method access
Previously, beginning the right operand with a decimal separator
produced a type error because the initial value of the right operand's
absolute value is null, and null has no methods.

Now, when the right operand's absolute value is null, the attempt to
access it will return a falsy value, and the decimal separator will be
concatenated to the right operand as intended.